refactor(eval): rename generic strategy solver to format strategy solver
